+++ Tests/Pcap++Test/main.cpp
@@ -4640,7 +4640,7 @@ void saveStringToFile(std::string& str, std::string fileName)
 	outfile.close();
 }
 
-void tcpReassemblyMsgReadyCallback(int sideIndex, const TcpStreamData& tcpData, void* userCookie)
+void tcpReassemblyMsgReadyCallback(int8_t sideIndex, const TcpStreamData& tcpData, void* userCookie)
 {
 	TcpReassemblyMultipleConnStats::Stats &stats = ((TcpReassemblyMultipleConnStats*)userCookie)->stats;
